Support drag n drop and closure based children (#10)
Adds dragDataSource and onDrop modifiers to OutlineView, which provide the ability to drag items from the OutlineView and receive dragged items. Dragging items with the dragDataSource requires converting an OutlineView element into a an NSPasteboardItem to initiate a drag. Receiving items with onDrop requires an instance of a type conforming to DropReceiver, which acts as a delegate to determine the behavior during the drag and once it is dropped onto the OutlineView.

Also adds new initializers to allow children of OutlineView data elements to be provided using a closure.

A new example app has been added, showing the usage of all the newly added features.